Rename RAG pipeline scripts
